# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /applications:
    post:
      operationId: ApplicationsController_create
      summary: Откликнуться на пост
      description: Один отклик на пост. Нельзя откликнуться на свой или архивный пост.
        ACCEPTED создаёт задачу. Создаёт сообщение «Новый отклик» в диалоге с
        владельцем поста. Владельцу поста отправляется email на Account OWNER.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "400":
          description: Недопустимый отклик
        "404":
          description: Пост не найден
        "409":
          description: Уже откликались на этот пост
      tags: &a14
        - applications
      security: &a15
        - access-token: []
  /applications/mine:
    get:
      operationId: ApplicationsController_listMine
      summary: Мои отклики
      description: "Отклики активного профиля с краткой информацией о посте. Фильтры:
        `type` (CREATOR / COMPANY), `status`, `createdDate` (YYYY-MM-DD, UTC).
        Поиск: `q` — по названию поста или названию компании."
      parameters:
        - name: postId
          required: false
          in: query
          description: Фильтр по посту (для входящих откликов)
          schema:
            format: uuid
            type: string
        - name: status
          required: false
          in: query
          schema:
            enum:
              - NEW
              - VIEWED
              - ACCEPTED
              - REJECTED
              - WITHDRAWN
            type: string
        - name: q
          required: false
          in: query
          description: "Поиск: для /applications/mine — по названию поста или
            компании-автора; для /applications/incoming — по названию поста"
          schema:
            example: реклама
            type: string
        - name: type
          required: false
          in: query
          description: Фильтр по типу поста (CREATOR / COMPANY)
          schema:
            enum:
              - CREATOR
              - COMPANY
            type: string
        - name: createdDate
          required: false
          in: query
          description: Фильтр по дате создания отклика (календарный день, UTC)
          schema:
            format: date
            example: 2026-06-14
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
      responses:
        "200":
          description: Список откликов с пагинацией
      tags: *a14
      security: *a15
  /applications/incoming:
    get:
      operationId: ApplicationsController_listIncoming
      summary: Входящие отклики на мои посты
      description: "Все отклики на посты активного профиля (владелец). Фильтры:
        `postId`, `status`, `createdDate` (YYYY-MM-DD, UTC), `type` (CREATOR /
        COMPANY). Поиск: `q` — по названию поста. Содержит данные поста и
        соискателя (applicant)."
      parameters:
        - name: postId
          required: false
          in: query
          description: Фильтр по посту (для входящих откликов)
          schema:
            format: uuid
            type: string
        - name: status
          required: false
          in: query
          schema:
            enum:
              - NEW
              - VIEWED
              - ACCEPTED
              - REJECTED
              - WITHDRAWN
            type: string
        - name: q
          required: false
          in: query
          description: "Поиск: для /applications/mine — по названию поста или
            компании-автора; для /applications/incoming — по названию поста"
          schema:
            example: реклама
            type: string
        - name: type
          required: false
          in: query
          description: Фильтр по типу поста (CREATOR / COMPANY)
          schema:
            enum:
              - CREATOR
              - COMPANY
            type: string
        - name: createdDate
          required: false
          in: query
          description: Фильтр по дате создания отклика (календарный день, UTC)
          schema:
            format: date
            example: 2026-06-14
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
      responses:
        "200":
          description: Список входящих откликов с пагинацией
      tags: *a14
      security: *a15
  /applications/{id}/withdraw:
    patch:
      operationId: ApplicationsController_withdraw
      summary: Отозвать отклик
      description: Только соискатель. Статусы NEW и VIEWED → WITHDRAWN.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "400":
          description: Нельзя отозвать в текущем статусе
        "403":
          description: Не ваш отклик
        "404":
          description: Отклик не найден
      tags: *a14
      security: *a15
  /applications/{id}/status:
    patch:
      operationId: ApplicationsController_updateStatus
      summary: Изменить статус отклика
      description: Только владелец поста. NEW → VIEWED/ACCEPTED/REJECTED; VIEWED →
        ACCEPTED/REJECTED.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "400":
          description: Недопустимый переход статуса
        "403":
          description: Не владелец поста
        "404":
          description: Отклик не найден
      tags: *a14
      security: *a15
...
info:
  title: Project Back API
  description: >-
    # Project Back API


    Backend для marketplace creator ↔ company. Документация для фронтенда и
    AI-ассистентов.


    ## Модель данных


    | Сущность | Что это |

    |----------|---------|

    | **Account** | Login: email + password. Один человек = один Account. |

    | **User** | Публичный профиль: CREATOR или COMPANY. При регистрации
    создаётся один User на Account. |

    | **Membership** | Доступ Account к User: OWNER / ADMIN / EDITOR / VIEWER. |

    | **Invite** | Приглашение другому человеку управлять **существующим**
    профилем. |


    **Активный профиль** — тот, чей `userId` в JWT (`sub`). Переключение: `POST
    /auth/switch-profile`.


    ## Cookies и CORS


    - После login/register/refresh выставляются httpOnly cookies:
    `access-token`, `refresh-token`.

    - Все запросы с `credentials: 'include'`.

    - CORS origin: `CORS_ORIGIN` (обычно http://localhost:3000).


    ## Сценарии


    ### Первый вход

    1. `POST /auth/register/creator` или `/auth/register/company` — создаёт
    Account + первый профиль.

    2. `POST /auth/login` — вход по email/password.


    ### Совместное управление (менеджер)

    1. Владелец: `POST /auth/invites` — email менеджера + `userId`
    **существующего** профиля + роль.

    2. Менеджер: регистрация/логин **своим** email → `POST
    /auth/invites/accept`.

    3. Менеджер: `GET /auth/profiles` → `POST /auth/switch-profile`.


...

...

...


...

...

...


...

...

...

...

...
servers:
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: access-token
  schemas:
...
    ApplicationCreatorProfileDto:
      type: object
      properties:
        name:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
      required:
        - name
        - lastName
    ApplicationCompanyProfileDto:
      type: object
      properties:
        companyName:
          type: string
          example: Acme Inc.
      required:
        - companyName
    ApplicationOwnerDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        avatar:
          type: string
          example: https://example.com/avatar.jpg
        creatorProfile:
          $ref: "#/components/schemas/ApplicationCreatorProfileDto"
        companyProfile:
          $ref: "#/components/schemas/ApplicationCompanyProfileDto"
      required:
        - id
    PostMediaDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          example: https://s3.ru-7.storage.selcloud.ru/project-media/posts/uuid/file.jpg
        key:
          type: string
          example: posts/uuid/file.jpg
        size:
          type: string
          example: "12345"
        mimeType:
          type: string
          example: image/jpeg
      required:
        - id
        - url
        - key
        - size
        - mimeType
...
    CreateApplicationDto:
      type: object
      properties:
        postId:
          type: string
          format: uuid
        message:
          type: string
          example: Готов обсудить сотрудничество. Опыт 3 года.
          minLength: 1
          maxLength: 2000
      required:
        - postId
        - message
    ApplicationPostSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Заголовок поста
        type:
          type: string
          enum:
            - CREATOR
            - COMPANY
        ownerId:
          type: string
          format: uuid
        owner:
          $ref: "#/components/schemas/ApplicationOwnerDto"
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMediaDto"
      required:
        - id
        - title
        - type
        - ownerId
    ApplicationApplicantDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - CREATOR
            - COMPANY
        avatar:
          type: string
          nullable: true
        name:
          type: string
          example: John
          description: Для CREATOR
        lastName:
          type: string
          example: Doe
          description: Для CREATOR
        companyName:
          type: string
          example: Acme Inc.
          description: Для COMPANY
      required:
        - id
        - role
    ApplicationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
          example: Готов обсудить сотрудничество
        status:
          type: string
          enum:
            - NEW
            - VIEWED
            - ACCEPTED
            - REJECTED
            - WITHDRAWN
          example: NEW
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        post:
          $ref: "#/components/schemas/ApplicationPostSummaryDto"
        applicant:
          $ref: "#/components/schemas/ApplicationApplicantDto"
      required:
        - id
        - message
        - status
        - createdAt
        - updatedAt
    UpdateApplicationStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - VIEWED
            - ACCEPTED
            - REJECTED
          example: ACCEPTED
          description: VIEWED, ACCEPTED или REJECTED
      required:
        - status
...